Redone large portions in Canvas3D class
stuff

+++ src/org/chaos/client/cache/definition/ItemDefinition.java
@@ -5751,7 +5751,7 @@ public final class ItemDefinition {
 		int l1 = Canvas3D.centerY;
 		int ai[] = Canvas3D.lineOffsets;
 		int ai1[] = Canvas2D.pixels;
-		float depth[] = Canvas2D.depthBuffer;
+		int depth[] = Canvas2D.depthBuffer;
 		int i2 = Canvas2D.width;
 		int j2 = Canvas2D.height;
 		int k2 = Canvas2D.topX;
@@ -5759,7 +5759,7 @@ public final class ItemDefinition {
 		int i3 = Canvas2D.topY;
 		int j3 = Canvas2D.bottomY;
 		Canvas3D.notTextured = false;
-		Canvas2D.initDrawingArea(32, 32, sprite2.myPixels, new float[32*32]);
+		Canvas2D.initDrawingArea(32, 32, sprite2.myPixels, new int[32*32]);
 		Canvas2D.drawPixels(32, 0, 0, 0, 32);
 		Canvas3D.method364();
 		int k3 = definition.modelZoom;
@@ -5925,7 +5925,7 @@ public final class ItemDefinition {
 		int l1 = Canvas3D.centerY;
 		int ai[] = Canvas3D.lineOffsets;
 		int ai1[] = Canvas2D.pixels;
-		float depth[] = Canvas2D.depthBuffer;
+		int depth[] = Canvas2D.depthBuffer;
 		int i2 = Canvas2D.width;
 		int j2 = Canvas2D.height;
 		int k2 = Canvas2D.topX;
@@ -5933,7 +5933,7 @@ public final class ItemDefinition {
 		int i3 = Canvas2D.topY;
 		int j3 = Canvas2D.bottomY;
 		Canvas3D.notTextured = false;
-		Canvas2D.initDrawingArea(32, 32, sprite2.myPixels, new float[32*32]);
+		Canvas2D.initDrawingArea(32, 32, sprite2.myPixels, new int[32*32]);
 		Canvas2D.drawPixels(32, 0, 0, 0, 32);
 		Canvas3D.method364();
 		int k3 = definition.modelZoom;
